# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar GreenPay API
  version: 1.0.0
  description: |
    OpenAPI 3.0 documentation for the backend REST API used by Stellar GreenPay.
    The API exposes project discovery, donations, profiles, leaderboard, notifications,
    ratings, updates, jobs, stats, and admin management endpoints.
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Local development server
  - url: https://api.stellargreenpay.com
    description: Production server
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ProjectIdParam:
      name: projectId
      in: path
      required: true
      description: Project identifier.
      schema:
        type: string
...
    PublicKeyParam:
      name: publicKey
      in: path
      required: true
      description: Stellar public key.
      schema:
        type: string
        pattern: '^G[A-Z0-9]{55}$'
...
  responses:
    BadRequest:
      description: The request payload or query parameters are invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid request payload
      required:
        - error
...
    Donation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        donorAddress:
          type: string
        amount:
          type: string
        amountXLM:
          type: string
          nullable: true
        currency:
          type: string
        message:
          type: string
          nullable: true
        transactionHash:
          type: string
        createdAt:
          type: string
          format: date-time
...
    CreateDonationRequest:
      type: object
      required: [projectId, donorAddress, amountXLM, transactionHash]
      properties:
        projectId:
          type: string
          format: uuid
        donorAddress:
          type: string
          pattern: '^G[A-Z0-9]{55}$'
        amountXLM:
          type: number
          minimum: 0.0000001
        amount:
          type: number
        currency:
          type: string
          default: XLM
        message:
          type: string
          maxLength: 100
        transactionHash:
          type: string
          pattern: '^[a-fA-F0-9]{64}$'
...
paths:
...
  /api/donations:
    post:
      tags: [Donations]
      summary: Record a donation
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDonationRequest'
      responses:
        '201':
          description: Donation recorded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Donation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/donations/project/{projectId}/messages:
    get:
      tags: [Donations]
      summary: List donation messages for a project
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Donation messages.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Donation'
  /api/donations/project/{projectId}:
    get:
      tags: [Donations]
      summary: List donations for a project
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Project donation history.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Donation'
                  nextCursor:
                    type: string
                    nullable: true
  /api/donations/donor/{publicKey}:
    get:
      tags: [Donations]
      summary: List donations by donor
      parameters:
        - $ref: '#/components/parameters/PublicKeyParam'
      responses:
        '200':
          description: Donor donation history.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Donation'
        '400':
          $ref: '#/components/responses/BadRequest'
...